requirements.txt
click==8.1.7
jsonpickle==3.0.2
matplotlib==3.7.3
numpy==1.24.4
pandas==2.0.3
prettytable==3.9.0
python-dateutil==2.8.2
schedule==1.2.0